requirements-dev.txt
coverage>=4.4.0
matplotlib>=2.1
seaborn
black[jupyter]~=24.1
pylint>=2.15.0
pylatexenc>=1.4
stestr>=2.0.0
ddt>=1.2.0,!=1.4.0
reno>=3.4.0
Sphinx>=5.0
sphinx-design>=0.4.0
sphinxcontrib-spelling
jupyter-sphinx
discover
qiskit-aer>=0.11.2
mypy>=0.981
mypy-extensions>=0.4.3
nbsphinx
qiskit_sphinx_theme~=1.16.0
qiskit-aer>=0.16
qiskit-ibm-runtime>=0.38
